# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: ForenseID API
  version: 1.0.0
  description: |
    🛡️ **ForenseID** - Sistema Profesional de Análisis Forense de Identidad
    
    Validación e inteligencia artificial para documentos de identidad mexicanos e internacionales.
    
    ## Características
    - 🔍 **OCR Avanzado**: Extracción de datos de INE, Pasaportes, Cédulas
    - 🤖 **Detección de Alteraciones**: Análisis EXIF, metadatos, patrones digitales
    - 👤 **Biometría Facial**: Detección de rostro y confianza de reconocimiento
    - ✅ **Validación de CURP/RFC**: Validación matemática de dígitos verificadores
    - 🌐 **Cross-Validation**: Coincidencia de datos entre múltiples fuentes
    
    ## Autenticación
    Usa tu **API Key** en uno de estos formatos:
    ```
    Authorization: Bearer your_api_key_here
    ?api_key=your_api_key_here
    {"api_key": "your_api_key_here"}
    ```
    
    **Demo Key**: `forenseid_demo_key_2026`
  
  contact:
    name: Soporte ForenseID
    email: support@forenseid.mx
    url: https://forenseid.mx
  
  license:
    name: Proprietary
    url: https://forenseid.mx/license
servers:
  - url: http://localhost:4000/api/v1
    description: Desarrollo Local
  - url: https://forenseid-api.vercel.app/api/v1
    description: Producción (Vercel)
...
paths:
...
  /sessions:
    post:
      tags:
        - Sesiones
      summary: Crear nueva sesión de verificación
      description: Inicia una nueva sesión de verificación de identidad
      operationId: createSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  example: "user_12345"
                durationMinutes:
                  type: integer
                  example: 30
              required:
                - userId
      responses:
        '201':
          description: Sesión creada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                  code:
                    type: string
                    example: "SESSION_CREATED"
                  data:
                    $ref: '#/components/schemas/SessionRecord'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    get:
      tags:
        - Sesiones
      summary: Listar sesiones
      description: Obtener lista de sesiones con filtros opcionales
      operationId: listSessions
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, EXPIRED, CANCELLED]
      responses:
        '200':
          description: Lista de sesiones
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionRecord'
                  details:
                    type: object
                    properties:
                      pagination:
                        type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /sessions/{id}:
    get:
      tags:
        - Sesiones
      summary: Obtener detalles de sesión
      description: Obtener información detallada de una sesión específica
      operationId: getSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalles de la sesión
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                  data:
                    $ref: '#/components/schemas/SessionRecord'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags:
        - Sesiones
      summary: Actualizar sesión
      description: Actualizar estado y datos de una sesión
      operationId: updateSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [ACTIVE, COMPLETED, EXPIRED, CANCELLED]
                documentId:
                  type: string
                result:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Sesión actualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
                  data:
                    $ref: '#/components/schemas/SessionRecord'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Sesiones
      summary: Cancelar sesión
      description: Cancela una sesión existente
      operationId: deleteSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sesión cancelada
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success]
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API Key como Bearer token
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: API Key como parámetro de query
  schemas:
...
    SessionRecord:
      type: object
      properties:
        id:
          type: string
          example: "cmno3nqop0004f5w7kvknhenl"
        createdAt:
          type: string
          format: date-time
        userId:
          type: string
          example: "user_12345"
        status:
          type: string
          enum: [ACTIVE, COMPLETED, EXPIRED, CANCELLED]
          example: "ACTIVE"
        documentId:
          type: string
          nullable: true
        verificationData:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        result:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
  responses:
    BadRequestError:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [error]
              code:
                type: string
                example: INVALID_REQUEST
              message:
                type: string
              request_id:
                type: string
    UnauthorizedError:
      description: Requiere autenticación válida
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [error]
              code:
                type: string
                example: INVALID_API_KEY
              message:
                type: string
              request_id:
                type: string
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [error]
              code:
                type: string
                example: NOT_FOUND
              message:
                type: string
              request_id:
                type: string
...